+++ sydtest/CHANGELOG.md
@@ -9,6 +9,7 @@
 * Changed the internals of `TestDefM` to use `ReaderT` and a strict `WriterT` instead of `RWST`.
 * Renamed `wrapRWST` to `wrapForest`.
 * Fixed the property label output to use the right total.
+* Moved `filterTestForest` and `randomiseTestForest` from `Test.Syd.TestDefM` to `Test.Syd.SpecDef`, where `SpecDefForest` is defined.
 
 ### Added